Replace frontend environment variables on runtime
When Next.js performs an export, it inlines all environment
variables. For docker, this means environment variables
can only be set as build arguments and not as runtime
environment variables.
To fix this, docker now builds with baked-in env values
and on init the backend replaces those values with the
real runtime environment variables.