Fix registration token vulnerability
If you try to register with the token set as the nil value of UUID then
the DB query sees the ID field as nil, so the WHERE clause is dropped.
As a result, it always grabs the first token in `active_tokens` and
you can register for whomever that token was created.